Intuitionize m1lgs and a few other theorems (#4733)
* Add m1lgs to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in several places
but is basically the set.mm proof.

* copy mullidd from set.mm to iset.mm

* copy elfzd from set.mm to iset.mm

* Add lgseisenlem1 and lgseisenlem2 to iset.mm

Stated as in set.mm.  The proofs need intuitionizing at a number of
steps but are basically the set.mm proofs.

* Add lgseisen to mmil.html

Also includes lgseisenlem3 and lgseisenlem4 .

* Add lgsquad to mmil.html missing theorems list

* add lgsquad and lgsquad2 to mmil.html

* add 2lgs to mmil.html

* Add sqdividap to iset.mm

This is sqdivid from set.mm with not equal changed to apart.  The proof
is the set.mm proof with not equal changed to apart at each step.

* Add mulsubdivbinom2ap to iset.mm

This is mulsubdivbinom2 from set.mm with not equal changed to apart.
The proof is the set.mm proof with not equal changed to apart at each
step.

* Add 2lgsoddprmlem1 to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* Add dvdsaddre2b to iset.mm

Stated as in set.mm.  The proof is based on a portion of the set.mm
proof but takes advantage of the fact that both sides of the
biconditional constrain B to be an integer.

* Add 2lgsoddprmlem2 to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in a few places but
is basically the set.mm proof.

* Add 2lgsoddprmlem3a to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* Add 2lgsoddprmlem3b to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* Rename mulid2i to mullidi in iset.mm

This is to match current set.mm naming.

* Add 2lgsoddprmlem3c to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* Add 2lgsoddprmlem3d to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* copy 2lgsoddprmlem3 from set.mm to iset.mm

* copy 2lgsoddprmlem4 from set.mm to iset.mm

* Add 2lgsoddprm to mmil.html

Since we won't be able to prove it now, adjust comments which refer to
it.