Intuitionize left modules from df-lmod to lmod0vid (#4737)
* Add LMod and .sf to iset.mm

This is the syntax, df-lmod , and df-scaf .  Copied without change from
set.mm.

* Add islmod to iset.mm

Stated as in set.mm.  The proof needs intuitionizing at quite a few
steps but the overall structure is the same as the set.mm proof.

* copy lmodlema from set.mm to iset.mm

* copy islmodd from set.mm to iset.mm

* copy lmodgrp from set.mm to iset.mm

* copy lmodring from set.mm to iset.mm

* copy lmodfgrp from set.mm to iset.mm

* copy lmodgrpd from set.mm to iset.mm

* Add lmodbn0 to iset.mm

Copied from set.mm.  The only change is to expand the comment regarding
the base set being inhabited.

* copy lmodacl from set.mm to iset.mm

* copy lmodmcl from set.mm to iset.mm

* Add lmodsn0 to iset.mm

Copied from set.mm.  The only change is to the comment, to also mention
the set of scalars being inhabited.

* Add lmodvacl , lmodass , lmodlcan to iset.mm

Copied without change from set.mm.

* Add lmodvscl to iset.mm

Copied from set.mm.  The only change is in the comment, to remove a
reference to a theorem not in iset.mm.

* Add scaffvalg to iset.mm

This is scaffval from set.mm with an additional set existence condition.
The proof is based on a portion of the set.mm proof.

* Add scafvalg to iset.mm

This is scafval from set.mm with an additional set existence condition.
The proof is similar to the set.mm proof but needs intuitionizing throughout.

* Add scafeqg to iset.mm

This is scafeq from set.mm with an additional set existence condition.
The proof needs intuitionizing but is basically the set.mm proof.

* Add scaffng to iset.mm

This is scaffn from set.mm with a set existence condition added.  The
proof needs a fair bit of intuitionizing but is basically the set.mm
proof.

* Add lmodscaf to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* Add lmodvsdi to iset.mm

Copied from set.mm.  The only change is to the comment, to remove a
reference to a theorem we don't have.

* Add lmodvsdir to iset.mm

Copied from set.mm.  The only change is to the comment, to remove a
reference to a theorem we don't have.

* Add lmodvsass to iset.mm

Copied from set.mm.  The only change is to the comment, to remove a
reference to a theorem we do not have.

* copy lmod0cl from set.mm to iset.mm

* copy lmod1cl from set.mm to iset.mm

* Add lmodvs1 to iset.mm

Copied from set.mm.  The only change is to the comment, to remove a
reference to a theorem we don't have.

* Add lmod0vcl to iset.mm

Copied from set.mm.  The only change is to the comment, to remove a
reference to a theorem we don't have.

* Add lmod0vlid to iset.mm

Copied from set.mm.  The only change is to the comment, to remove a
reference to a theorem we don't have.

* Add lmod0vrid to iset.mm

Copied from set.mm.  The only change is to the comment, to remove a
reference to a theorem we don't have.

* copy lmod0vid from set.mm to iset.mm

* Adjust comment for lmod0cl and lmod1cl

The reference is to Base ` F not Base ` W which is the reason for the
change.  Changed in set.mm and iset.mm.