Check bibliographic references in set.mm and iset.mm (#4739)
* Add theorem list check

This will run write theorem_list on all pull requests and fail the
checks if it produces any errors.

It checks set.mm and iset.mm.

* Add [Halmos] to mmset.raw.html

It had been in mmhil.html but because there is a reference in set.mm, it
needs to be in mmset.raw.html as well.

* Add bibliography entries to mmil.raw.html

These are copied from mmset.raw.html and are for entries with references
in iset.mm.

Also, alphabetize some of the existing entries.

* Expand check-theorem-list comment