Intuitionize left modules from lmod0vs to rmodislmod (#4744)
* Add lmod0vs to iset.mm

Copied from set.mm.  The only change is to the comment, to remove a
reference to a theorem we don't have.

* Add lmodvs0 to iset.mm

Copied from set.mm.  The only change is to the comment, to remove a
reference to a theorem we don't have.

* copy mulgnn0cld from set.mm to iset.mm

* copy lmodvsmmulgdi from set.mm to iset.mm

* Add lmodfopnelem1 to iset.mm

Stated as in set.mm.  The proof needs intuitionization but is basically
the set.mm proof.

* copy lmodfopnelem2 from set.mm to iset.mm

* Add lmodfopne to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy lcomf from set.mm to iset.mm

* add lcomfsupp to mmil.html

* copy lmodvnegcl from set.mm to iset.mm

* copy lmodvnegid from set.mm to iset.mm

* copy lmodvneg1 from set.mm to iset.mm

* copy lmodvsneg from set.mm to iset.mm

* Add lmodvsubcl to iset.mm

Copied from set.mm.  The only change is to the comment, to remove a
reference to a theorem iset.mm doesn't have.

* copy lmodcom from set.mm to iset.mm

* copy lmodabl from set.mm to iset.mm

* copy lmodcmn from set.mm to iset.mm

* copy lmodnegadd from set.mm to iset.mm

* copy lmod4 from set.mm to iset.mm

* Add lmodvsubadd to iset.mm

Copied from set.mm.  The only change is the comment, to remove a
reference to a theorem that iset.mm doesn't have.

* copy lmodvaddsub4 from set.mm to iset.mm

* Add lmodvpncan to iset.mm

Copied from set.mm.  The only change is to the comment, to remove a
reference to a theorem that iset.mm doesn't have.

* Add lmodvnpcan to iset.mm

Copied from set.mm.  The only change is to the comment, to remove a
reference to a theorem that iset.mm doesn't have.

* Add lmodvsubval2 to iset.mm

Copied from set.mm.  The only change is to the comment, to remove a
reference to a theorem that iset.mm doesn't have.

* copy lmodsubvs from set.mm to iset.mm

* Add missing changes-set.txt entries

These are for rng->ring in labels, pulled from a 12-Feb-2025 git log.

* Rename rngo2times to ringo2times in iset.mm

This is to match set.mm.

* Rename rngnegr to ringnegr in iset.mm

This is to match set.mm

* Rename rngsubdir to rngsubdir in iset.mm

This is to match set.mm.

* Add lmodsubdi to iset.mm

Copied from set.mm.  The only change is to the comment, to remove a
reference to a theorem iset.mm doesn't have.

* Add lmodsubdir to iset.mm

Copied from set.mm.  The only change is to the comment, to remove a
reference to a theorem not in iset.mm.

* Add lmodsubeq0 to iset.mm

Copied from set.mm.  The only change is to remove a reference to a
theorem not in iset.mm.

* Add lmodsubid to iset.mm

Copied from set.mm.  The only change is to the comment, to remove a
reference to a theorem not in iset.mm.

* add lmodvsghm to mmil.html

* Add lmodprop2d to iset.mm

Stated as in set.mm.  One part of the proof needs significant
intuitionizing but other than that, it is the set.mm proof.

* copy lmodpropd from set.mm to iset.mm

* add gsumvsmul to mmil.html

* add mptscmfsupp0 and mptscmfsuppd to mmil.html

* copy ralrot3 from set.mm to iset.mm

* add rspn0 to mmil.html

* Add rmodislmodlem to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in various places
but is basically the set.mm proof.

* Add rmodislmod to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in a number of
places but is basically the set.mm proof.