Add LSubSp to iset.mm (#4758)
* Add LSubSp to iset.mm

This is the syntax and section header.  Copied without change from
set.mm.

* Add df-lssm to iset.mm

This is df-lss from set.mm with not empty changed to inhabited.

* Add lsssetm to iset.mm

This is lssset from set.mm with not empty changed to inhabited.  The
proof is the set.mm proof with changes to reflect that change.

* Add islssm to iset.mm

This is islss from set.mm with a few changes.  The proof is similar to
the set.mm proof.

* Add islssmd to iset.mm

This is similar to islssd from set.mm, analogous to islssm compared with
islss .

* Add lssssg to iset.mm

This is lssss from set.mm with a set existence condition added.  The
proof is similar to the set.mm proof.

* Add lsselg to iset.mm

This is lssel from set.mm with a set existence condition added.  The
proof is similar to the set.mm proof.

* Adjust lmodbn0 comment

There is more direct theorem which shows that a left module's base set
is inhabited.

* Add lss1 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lssuni to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* add lssn0 to mmil.html

* Add function value theorems to mmil.html

This is funfv , funfv2 , funfv2f , fvun , dffv2 , and fvco4i .
The information is mostly taken from git logs.

* Add lssclg to iset.mm

This is lsscl from set.mm with a set existence condition added.  The
proof is similar to the set.mm proof.

* Add lssvsubcl to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lssvancl1 to iset.mm

Stated as in set.mm.  The proof needs a small amount of inuitionizing
but is mostly the set.mm proof.  Remove references in comment to
theorems that iset.mm does not have.

* Add lssvancl2 to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* Add lss0cl to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add 00lss to mmil.html

Link to lss0cl

* Add lsssn0 to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is basically
the set.mm proof.

* Add lss0ss to iset.mm

Copied from set.mm.  The only change is to the comment, to remove a
reference to a theorem we don't have.

* Add lssle0 to iset.mm

Copied from set.mm.  The only change is to the comment, to remove a
reference to a theorem we don't have.

* add lssne0 to mmil.html

* copy lssvneln0 from set.mm to iset.mm

* copy lssneln0 from set.mm to iset.mm

* add lssssr to mmil.html

* Add lssvacl to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* Add lssvscl to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
basically the set.mm proof.

* Add lssvnegcl to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
basically the set.mm proof.

Also, minimize the set.mm proof via scripts/minimize.

* Add lsssubg to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
basically the set.mm proof.

* copy lsssssubg from set.mm to iset.mm

* Add ressscag to iset.mm

This is resssca from set.mm with a set existence condition added.  The
proof is similar to the set.mm proof.

* Add ressvscag to iset.mm

This is ressvsca from set.mm with an additional set existence condition.
The proof is similar to the set.mm proof.

* Add ressipg to iset.mm

This is ressip in set.mm with a set existence condition added.  The
proof is similar to the set.mm proof.

* Add islss3 to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in a variety of
places but is basically the set.mm proof.

* copy lsslmod from set.mm to iset.mm

* Add lsslss to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add islss4 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Rename a1tru to trud to iset.mm

This matches the current naming in set.mm.

* Revise cbvrexfw comment in iset.mm.

The previous wording had the wrong axiom numbering.  There's a sentence
in the cbvralfw comment which also applies here, so copy that over.

* Add cbvrexw to iset.mm

Copied from set.mm.  The only change is in the comment, to reflect
current and future axiom usage.

* copy elabrexg from set.mm to iset.mm

* Revise lmodsn0 comment in iset.mm.

Give a more concise way of showing that the set of scalars is inhabited.

* Add lss1d to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in various places,
tricky at times, but is basically the set.mm proof.

* Add lssintclm to iset.mm

This is lssintcl from set.mm with nonempty changed to inhabited.  The
proof needs intuitionizing in various ways but is basically the set.mm
proof.

* Add lssincl to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
basically the set.mm proof.  Also, expand comments of prnz and prnzg to
mention prm and prmg .

* Add lssmre and lssacs to mmil.html

* add prdsvscacl to mmil.html

* add prdslmodd to mmil.html

* add pwslmod to mmil.html

* Move elabrexg from mathbox to main in set.mm

This is because it is in main in iset.mm.

* Revise df-lssm comment.

As suggested on the pull request, elaborate a bit more about what it
does.

* Correct lssintclm comment

nonempty -> inhabited