Intuitionize LSpan through lsppropd (#4768)
* Add LSpan to iset.mm

Copied without change from set.mm.  This is the syntax and df-lsp .

* Remove $d ph x $. from rabeqdv in iset.mm

This was removed from set.mm a while ago and the proof does not need it.

* Add lspfval to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lspf to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lspval to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lspcl to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy lspsncl from set.mm to iset.mm

* copy lspprcl from set.mm to iset.mm

* copy lsptpcl from set.mm to iset.mm

* copy lspsnsubg from set.mm to iset.mm

* add 00lsp to mmil.html

* Add lspid to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* Add lspssv to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* Add lspss to iset.mm

Copied from set.mm with only a small change to the comment.

* Add lspssid to iset.mm

Copied from set.mm with only a small change to the comment.

* copy lspidm from set.mm to iset.mm

* copy lspun from set.mm to iset.mm

* Add lspssp to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* add mrclsp to mmil.html

* Add lspsnss to iset.mm

Copied from set.mm with just one small change to the comment.

* Add lspsnel3 to iset.mm

Copied from set.mm with just one small change to the comment.

* copy lspprss from set.mm to iset.mm

* Add lspsnid to iset.mm

Copied from set.mm with only one small change to the comment.

* Add lspsnel6 to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* copy lspsnel5 from set.mm to iset.mm

* Add lspsnel5a to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* copy lspprid1 from set.mm to iset.mm

* copy lspprid2 from set.mm to iset.mm

* copy lspprvacl from set.mm to iset.mm

* Add lssats2 to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* Add lspsneli to iset.mm

Copied from set.mm with the only change being a small change to the
comment.

* copy lspsn from set.mm to iset.mm

* Add lspsnel to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in various ways
including using elrab3g instead of elrab3 , but is basically the set.mm
proof.

* copy lspsnvsi from set.mm to iset.mm

* Add lspsnss2 to iset.mm

Copied from set.mm with the only change being to the comment, to remove
a reference to a theorem we don't have yet.

* copy lspsnneg from set.mm to iset.mm

* copy lspsnsub from set.mm to iset.mm

* Add lspsn0 to iset.mm

Copied from set.mm with only one small change to the comment.

* copy lsp0 from set.mm to iset.mm

* Add lspuni0 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy lspun0 from set.mm to iset.mm

* copy lspsneq0 from set.mm to iset.mm

* copy lspsneq0b from set.mm to iset.mm

* copy lmodindp1 from set.mm to iset.mm

* Add lsslsp to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lss0v to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lsspropdg to iset.mm

This is lsspropd from set.mm with set existence conditions added.  The
proof is the set.mm proof with some intuitionizing.

* Add lsppropd to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* Wording fix to lspf comment

operator->function and subset->subspace.  Same change in set.mm and
iset.mm.