Intuitionize section "The subring algebra; ideals" (#4780)
* Start section Ideals in iset.mm

This is the syntaxes for subringAlg , ringLMod , RSpan , and LIdeal .
and the definitions df-sra , df-rgmod , df-lidl , and df-rsp .

Copied without change from set.mm.

* Add ovexd to mmil.html

* Add sraval to iset.mm

Stated as in set.mm.  Several parts of the proof need intuitionizing but
it is basically the set.mm proof.

* copy fv2prc from set.mm to iset.mm

* Add sralemg to iset.mm

This is sralem from set.mm with additional conditions and a decent
amount of intuitionizing in the proof.

* Add srabaseg to iset.mm

This is srabase from set.mm with additional conditions.  The proof is
very similar to the set.mm proof.

* Add sraaddgg to iset.mm

This is sraaddg from set.mm with an added condition.  The proof is
similar to the set.mm proof.

* Add sramulrg to iset.mm

This is sramulr from set.mm with a set existence condition added.  The
proof is very similar to the set.mm proof.

* Add srascag to iset.mm

This is srasca from set.mm with a set existence condition.  The proof
needs intuitionizing in a lot of places but is basically the set.mm
proof.

* Add sravscag to iset.mm

This is sravsca from set.mm with a set existence condition added.  The
proof needs a fair bit of intuitionizing in most places but is basically
the set.mm proof.

* Add sraipg and sratsetg to iset.mm

This is sraip and sratset from set.mm with a set existence condition
added.  The proofs need intuitionizing but are based on the set.mm
proofs.

* Add sratopng to iset.mm

This is sratopn from set.mm with a set existence condition added.
Although the proof is based on the set.mm proof, it ends up much longer
because of the needed intuitionizing.

* Add sradsg to iset.mm

This is srads from set.mm with a set existence condition added.  The
proof is very similar to the set.mm proof.

* Add sraring to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add sralmod to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is based
on the set.mm proof.

* Add sraex to iset.mm

In set.mm this would be handled by ovex .  Use it to shorten the sratopng proof.

* Add sralmod0g to iset.mm

This is sralmod0 from set.mm with a set existence condition added.  The
proof is similar to the set.mm proof.

* copy issubrgd from set.mm to iset.mm

* Add rlmfn to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is basically
the set.mm proof.

* Add rlmvalg to iset.mm

This is rlmval from set.mm with an added set existence condition.  The
proof is similar to a portion of the set.mm proof.

* Add lidlvalg to iset.mm

This is lidlval from set.mm with a set existence condition added.
The proof is similar to the set.mm proof.

* Add rspvalg to iset.mm

This is rspval from set.mm with a set existence condition added.  The
proof is similar to the set.mm proof.

* Add rlmval2 to mmil.html

* Add rlmbasg to iset.mm

This is rlmbas from set.mm with a set existence condition added.  The
proof is similar to the set.mm proof.

* Add rlmplusgg to iset.mm

This is rlmplusg from set.mm with an added set existence condition.  The
proof is similar to the set.mm proof.

* Add rlm0g to iset.mm

This is rlm0 from set.mm with a set existence condition added.  The
proof is similar to the set.mm proof.

* Add rlmsubg to iset.mm

This is rlmsub from set.mm with a set existence condition added.  The
proof is similar to the set.mm proof.

* Add rlmmulr to iset.mm

This is rlmmul from set.mm with a set existence condition added.  The
proof is similar to the set.mm proof.

* add rlmsca and rlmsca2 to mmil.html

* Add rlmvscag to iset.mm

This is rlmvsca from set.mm with an additional set existence condition.
The proof is similar to the set.mm proof.

* Add rlmtopng to iset.mm

This is rlmtopn from set.mm with an additional set existence condition.
The proof is similar to the set.mm proof.

* Add rlmdsg to iset.mm

This is rlmds from set.mm with a set existence condition added.  The
proof is similar to the set.mm proof.

* Add rlmlmod to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* add rlmlvec to mmil.html

* add rlmlsm to mmil.html

* Add rlmvnegg to iset.mm

This is rlmvneg with an additional set existence condition.  The proof
is similar to the set.mm proof.

* Add rlmscaf to mmil.html

* Add ixpsnbasval to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lidlmex to iset.mm

This is a case where we can get set existence from existing conditions.

* Add lidlss to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is based on the
set.mm proof.

* Add lidlssbas to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lidlbas to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* add islidl to mmil.html

* add rnglidlmcl to mmil.html

* add rngridlmcl to mmil.html

* Add lidl0cl to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lidlacl to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lidlnegcl to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lidlsubg to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is basically
the set.mm proof.

* copy lidlsubcl from set.mm to iset.mm

* Add lidlmcl to mmil.html

* add lidl1el to mmil.html

* Add dflidl2lem and dflidl2 to mmil.html

* Add lidl0 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lidl1 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* add lidlacs to mmil.html

* Add rspcl to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is basically
the set.mm proof.

* Add rspssid to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is basically
the set.mm proof.

* add rsp1 to mmil.html

* Add rsp0 to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is basically
the set.mm proof.

* Add rspssp to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is basically
the set.mm proof.

* add mrcrsp to mmil.html

* add lidlnz to mmil.html

* add drngnidl to mmil.html

* add lidlrsppropd to mmil.html

* Subsection "The subring algebra; ideals" split into two subsections