exponentiation injectivity
I prove cxp112d and cxp111d only to find shorter and
more fundamental proofs for rxp112d and rxp11d

Note: Unsure about naming. rxp11d is consistent with exp11d, so either
both should be *xp111d or cxp111d can be cxp11d. And rxp is an experiment
for "recxp", can be changed if someone is against this.