Intuitionize Z/nZ through znlidl (#4789)
* Add ZRHom , ZMod , and Z/nZ to iset.mm

This is the syntax.  Copied without change from set.mm.

Add df-chr to mmil.html.

* copy df-zrh and df-zlm from set.mm to iset.mm

* copy df-zn from set.mm to iset.mm

* add zrhval to mmil.html

* add zrhval2 and zrhmulg to mmil.html

* Add ZRHom theorems to mmil.html

This is zrhrhmb , zrhrhm , zrh1 , zrh0 , and zrhpropd .

* Add mulgex to iset.mm

A set existence theorem, not needed in set.mm because it has fvex and
the like.

* Add zlmval to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is basically
the set.mm proof.

* Add zlmbasg and zlmplusgg to iset.mm

This is zlmbas and zlmplusg from set.mm with set existence conditions
added.  Includes lemma zlmlemg , based on set.mm's zlmlem .  All three
proofs are similar to the corresponding set.mm proofs.

* Add zlmmulrg to iset.mm

This is zlmmulr from set.mm with a set existence condition added.  The
proof is similar to the set.mm proof.

* Add zlmsca to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is basically
the set.mm proof.

* Add zlmvscag to iset.mm

This is zlmvsca from set.mm with a set existence condition added.  The
proof is similar to the set.mm proof.

* add zlmlmod to mmil.html

* copy znlidl from set.mm to iset.mm

* Add imasgrp2 and qusgrp2 to iset.mm

Copy imasgrp2 from set.mm without change.

State qusgrp2 as in set.mm.  The proof needs a little bit of
intuitionizing but is very close to the set.mm proof.

* Add qusex to iset.mm

This is a set existence theorem, which would be handled by ovex in
set.mm.  The proof follows fairly readily from df-qus and existing set
existence theorems.

* add zncrng2 to mmil.html

* add znval to mmil.html

* Add eqgex to iset.mm

This is a set existence theorem.
The proof is taken from a portion of the releqgg proof.

* Add lspex to iset.mm

This is a set existence theorem.
The proof is fairly direct from lspfval .

* Add rspex to iset.mm

This is a set existence theorem.
The proof is direct from rspvalg , rlmfn , and lspex .