injectivity of exponentiation (#4793)
* exponentiation injectivity

I prove cxp112d and cxp111d only to find shorter and
more fundamental proofs for rxp112d and rxp11d

Note: Unsure about naming. rxp11d is consistent with exp11d, so either
both should be *xp111d or cxp111d can be cxp11d. And rxp is an experiment
for "recxp", can be changed if someone is against this.

* shorten imaexi

I couldn't get minimize to find this, but no OLD:

no information is lost since the proof was the same as for imaexg

(imaexi and imaex have the same proof but are different,
as seen by the new use of elexi)

* update changes-set.txt