Clarify ax-mulf comment
The ax-mulf comment is confusing.
Its text "is not a bona fide axiom for complex numbers"
makes it sound like this statement can't be an axiom or that
Metamath can't handle such axioms, which is obviously untrue.
Also, the cited paper doesn't justify this axiom.

This commit attempts to clarify the comment for ax-mulf
in the hopes of making it clearer.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>