Intuitionize 2Ideal and related topics (#4797)
* copy imasgrp from set.mm to iset.mm

* copy imasgrpf1 from set.mm to iset.mm

* copy imasring from set.mm to iset.mm

* copy imasringf1 from set.mm to iset.mm

The only change is to the comment, to remove a reference to a theorem we
don't have.

* Add qusring2 to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
basically the set.mm proof.

* Add 2Ideal to iset.mm

This is the syntax and df-2idl .  Copied without change from set.mm.

* Add lssex to iset.mm

This is a set existence theorem which set.mm would use fvex for.
The proof follows from df-lssm and mptfvex .

* Add lidlex to iset.mm

This is a set existence theorem which set.mm would use fvex for.
The proof follows from lssex and rlmfn .

* Add 2idlvalg to iset.mm

This is 2idlval from set.mm with an additional set existence condition.
The proof is taken from a portion of the set.mm with some additional
intuitionizing.

Also, add 2idlmex which is similar to lidlmex but for 2Ideal .  The
proof follows from relelfvdm and df-2idl .

* Rename islssm to islssmg in iset.mm

This is to free up the islssm name for a theorem which lacks the W e. X
condition.

* Add lssmex to iset.mm

This is a set existence theorem based on a set being inhabited.  It is
similar, including the proof, to 2idlmex .

* Add islssm to iset.mm

Similar to the previous islssm (now called islssmg ) except without the
added set existence condition.  Turns out both sides of the
biconditional imply that condition.

Mark islssmg as discouraged and update mmil.html.

* Add ressbasid to iset.mm

It may be hard to show that trivial restriction gives us the same
structure that we started with, but getting a structure with the same
base slot is easier.

* Add rlmscabas to iset.mm

This a portion of rlmsca which we can prove.

* Add islidlm to iset.mm

This is islidl from set.mm with nonempty changed to inhabited.  The
proof needs quite a bit of intuitionizing but is loosely based on the
set.mm proof.

* Rename grpridd to grprida in iset.mm

This matches a rename which was already done in set.mm.

* copy grplidd and grpridd from set.mm to iset.mm

* Add rnglidlmcl to iset.mm

Stated as in set.mm.  The proof is basically the set.mm proof with
nonempty changed to inhabited at a variety of steps.

* Revise rngridlmcl and lidlmcl in mmil.html

These are close so say so.  Revise one iset.mm comment which refers to lidlmcl .

* Copy ad5ant245 and others from set.mm to iset.mm

This is ad5ant245 , ad5ant234 , ad5ant235 , ad5ant123 , ad5ant124 ,
ad5ant125 , ad5ant134 , ad5ant135 , and ad5ant145 .

* Add dflidl2rng to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add Abel / CMnd / Ring / Rng theorems to iset.mm

This is ablcmnd , ringabld , ringcmnd , ringrng , and ringssrng .
Copied without change from set.mm.

* copy biadanid from set.mm to iset.mm

* copy dflidl2 from set.mm to iset.mm

* Add opprsubgg to iset.mm

This is opprsubg from set.mm with a set existence condition added.  The
proof needs a lot of intuitionizing but basically follows the set.mm
proof.

* Add isridl to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is based on the
set.mm proof.

* Add 2idlelb to iset.mm

Stated as in set.mm.  The proof is based on the set.mm proof but is
significantly longer to deal with set existence.  Revise comment to
remove reference to a theorem iset.mm doesn't have.