Intuitionize 2Ideal through 2idlelbas (#4809)
* Add 2idllidld to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is basically
the set.mm proof.

* Add 2idlridld to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is basically
the set.mm proof.

* Add opprrng to iset.mm

Stated as in set.mm.  The proof needs quite a bit of intuitionizing but
follows the structure of the set.mm proof.

* Add isridlrng to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is basically
the set.mm proof.

* copy df2idl2rng from set.mm to iset.mm

* copy df2idl2 from set.mm to iset.mm

* Add ridl0 to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* Add ridl1 to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* Add 2idl0 to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* Add 2idl1 to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* copy 2idlss from set.mm to iset.mm

* Add 2idlbas to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy 2idlelbas from set.mm to iset.mm

* add rng2idlsubrng to mmil.html

* Add rng2idlnsg to mmil.html

* add rng2idl0 to mmil.html

* Add more SubRng theorems to mmil.html

This is rng2idlsubgsubrng , rng2idlsubgnsg , and rng2idlsubg0 .

* add 2idlcpblrng to mmil.html

* Add 2idlcpbl to mmil.html

Remove reference to it in a comment.

* Add qus2idrng to mmil.html

* Add qus1 and friends to mmil.html

This is qus1 , qusring , qusrhm , and qusmul2

* Add crngridl to mmil.html

* add crng2idl to mmil.html

* add qusmulrng to mmil.html

* add quscrng to mmil.html